Move PluginProvider into App and extract existing providers into a plugin-able AppProviders component
